debug system reworked to be more global
