+++ rust/psibase_macros/psibase-macros-lib/src/service_tables_macro/mod.rs
@@ -4,7 +4,7 @@ use proc_macro2::TokenStream;
 use proc_macro_error::abort;
 use quote::quote;
 use std::{collections::HashMap, str::FromStr};
-use syn::{Ident, Item, ItemMod, Type};
+use syn::{parse_quote, Ident, Item, ItemMod, Type};
 use tables::{is_table_attr, process_service_tables};
 
 pub fn service_tables_macro_impl(_attr: TokenStream, item: TokenStream) -> TokenStream {
@@ -45,13 +45,30 @@ fn process_mod(psibase_mod: &proc_macro2::TokenStream, mut impl_mod: ItemMod) ->
             }
         }
 
+        let mut debug_msgs: Vec<String> = Vec::new();
         let mut processed_tables = Vec::new();
         for (tb_name, items_idxs) in table_structs.iter() {
-            let table_idx = process_service_tables(psibase_mod, tb_name, items, items_idxs);
+            let table_idx =
+                process_service_tables(psibase_mod, tb_name, items, items_idxs, &mut debug_msgs);
             processed_tables.push((tb_name, table_idx));
         }
 
+        let doc_attrs = debug_msgs.into_iter().map(|msg| {
+            quote! {
+                #[doc = #msg]
+            }
+        });
+
+        // Combine the struct and its documentation attributes into a single token stream
+        let output = parse_quote! {
+            #(#doc_attrs)*
+            struct macro_debug_msgs {}
+        };
+
+        items.push(output);
+
         // Validates table indexes
+        // TODO: re-enable this code
         processed_tables.sort_by_key(|t| t.1);
         // for (expected_idx, (table_struct, tb_index)) in processed_tables.iter().enumerate() {
         //     if *tb_index as usize != expected_idx {
