+++ services/user/Chainmail/service/src/tables/saved_messages.rs
@@ -10,6 +10,15 @@ pub mod tables {
     use psibase::ToSchema;
     use serde::{Deserialize, Serialize};
 
+    #[table(name = "InitTable", index = 0)]
+    #[derive(Serialize, Deserialize, ToSchema, Fracpack)]
+    pub struct InitRow {}
+
+    impl InitRow {
+        #[primary_key]
+        fn pk(&self) {}
+    }
+
     #[table(name = "SavedMessageTable", index = 1)]
     #[derive(Debug, Serialize, Deserialize, ToSchema, Fracpack, SimpleObject)]
     pub struct SavedMessage {
